renamed to trendpy2
