+++ trendpy2/__init__.py
