+++ trendpy2/methods.py
